requirements.txt
streamlit
firecrawl-py
langchain
langchain-community
faiss-cpu
openai==0.27.0
tiktoken